arreglando algunas cosas

+++ src/main/java/ar/edu/unq/epersgeist/persistencia/repositories/interfaces/MediumRepository.java
@@ -14,5 +14,5 @@ public interface MediumRepository {
     List<Medium> recuperarTodosLosEliminados();
     List<Espiritu> findEspiritusByMediumId(Long mediumId);
     void deleteAll();
-    Double laDistanciaA(Double longitud, Double latitud, Double idMediumSQL);
+    Double laDistanciaA(Double longitud, Double latitud, Long idMediumSQL);
 }
